requirements.txt
django==4.1.7
djangorestframework==3.14.0
django-filter==22.1
# psycopg2-binary==2.9.5

